+++ R/location_functions.R
@@ -80,12 +80,12 @@ decrypt_gps <- function(data, key, ignore = ":") {
 }
 
 deg2rad <- function(deg) {
-  check_arg(deg, "double")
+  check_arg(deg, "numeric")
   deg * pi / 180
 }
 
 rad2deg <- function(rad) {
-  check_arg(rad, "double")
+  check_arg(rad, "numeric")
   rad * 180 / pi
 }
 
@@ -110,11 +110,11 @@ rad2deg <- function(rad) {
 #' ord <- c(41.97861, -87.90472) # Chicago O'Hare International Airport
 #' haversine(fra[1], fra[2], ord[1], ord[2]) # 6971.059 km
 haversine <- function(lat1, lon1, lat2, lon2, r = 6371) {
-  check_arg(lat1, "double")
-  check_arg(lon1, "double")
-  check_arg(lat2, "double")
-  check_arg(lon2, "double")
-  check_arg(r, "double")
+  check_arg(lat1, "numeric")
+  check_arg(lon1, "numeric")
+  check_arg(lat2, "numeric")
+  check_arg(lon2, "numeric")
+  check_arg(r, "numeric")
 
   p <- pi / 180
   a <- 0.5 -
@@ -124,8 +124,8 @@ haversine <- function(lat1, lon1, lat2, lon2, r = 6371) {
 }
 
 location_variance <- function(lat, lon) {
-  check_arg(lat, "double")
-  check_arg(lon, "double")
+  check_arg(lat, "numeric")
+  check_arg(lon, "numeric")
 
   log((stats::sd(lat) * 2 + stats::sd(lon) * 2) + 1)
 }
